fix: cast captured attention to float32 before numpy conversion
FluxAttentionStore.add_attention stored captured tensors via
detach().cpu() with no dtype cast. On real GPU hardware DTYPE is
bfloat16, so every captured tensor would be bfloat16 -- and
cross_attention_map's tensor.numpy() call does not support bfloat16,
raising TypeError. Found during Task 6's code review by reproducing it
end-to-end (not theorized): this would have made every one of 105
prompts in the eventual GPU run fail with detected=False, producing
zero usable attention data from the whole run.

Fix: detach().float().cpu(), matching generate_anchor_images_sdxl.py's
existing pattern of casting to float32 before its own numpy path.

+++ ssa/anchor_set/flux_attention_capture.py
@@ -76,7 +76,7 @@ class FluxAttentionStore:
 
     def add_attention(self, layer_name: str, head_averaged: torch.Tensor) -> None:
         """`head_averaged`: (image_seq_len, n_targets) for one layer at the current step."""
-        self.step_store.setdefault(self.current_step, {})[layer_name] = head_averaged.detach().cpu()
+        self.step_store.setdefault(self.current_step, {})[layer_name] = head_averaged.detach().float().cpu()
 
     def step(self) -> None:
         self.current_step += 1
